2024/1/5 7:57PM
fixed the resume issue where it would not render on the mobile hosted website.

I converted the pdf to a png and hosted it with the typical imagehtml.

fixed the issue where the images on the homepage would change size when used on mobile

changed the images to be based on pure pixel values instead of view height.